backend/requirements.txt
flask
flask-cors
Pillow
numpy
torch
torchvision
opencv-python
segmentation-models-pytorch
requests
